refactor: Add unit test for read texture files.

+++ src/main/scala/moe/brianhsu/porting/live2d/demo/sprite/BackgroundSprite.scala
@@ -2,7 +2,7 @@ package moe.brianhsu.porting.live2d.demo.sprite
 
 import moe.brianhsu.live2d.boundary.gateway.renderer.DrawCanvasInfoReader
 import moe.brianhsu.live2d.enitiy.opengl.OpenGLBinding
-import moe.brianhsu.porting.live2d.renderer.opengl.TextureManager.TextureInfo
+import moe.brianhsu.live2d.usecase.renderer.texture.TextureInfo
 
 class BackgroundSprite(drawCanvasInfo: DrawCanvasInfoReader, textureInfo: TextureInfo, shader: SpriteShader)
                       (implicit private val gl: OpenGLBinding) extends LAppSprite(drawCanvasInfo, textureInfo, shader) {
@@ -11,7 +11,7 @@ class BackgroundSprite(drawCanvasInfo: DrawCanvasInfoReader, textureInfo: Textur
 
     val windowWidth = drawCanvasInfo.currentCanvasWidth
     val windowHeight = drawCanvasInfo.currentCanvasHeight
-    Position(0, 0, windowWidth, windowHeight)
+    Position(0, 0, windowWidth.toFloat, windowHeight.toFloat)
   }
 
 }
