+++ src/main/scala/moe/brianhsu/porting/live2d/demo/sprite/GearSprite.scala
@@ -2,7 +2,7 @@ package moe.brianhsu.porting.live2d.demo.sprite
 
 import moe.brianhsu.live2d.boundary.gateway.renderer.DrawCanvasInfoReader
 import moe.brianhsu.live2d.enitiy.opengl.OpenGLBinding
-import moe.brianhsu.porting.live2d.renderer.opengl.TextureManager.TextureInfo
+import moe.brianhsu.live2d.usecase.renderer.texture.TextureInfo
 
 class GearSprite(canvasInfo: DrawCanvasInfoReader, textureInfo: TextureInfo, shader: SpriteShader)
                 (implicit private val gl: OpenGLBinding) extends LAppSprite(canvasInfo, textureInfo, shader) {
@@ -11,8 +11,8 @@ class GearSprite(canvasInfo: DrawCanvasInfoReader, textureInfo: TextureInfo, sha
     val windowHeight = canvasInfo.currentCanvasHeight
 
     Position(
-      windowWidth - textureInfo.width,
-      windowHeight - textureInfo.height,
+      windowWidth - textureInfo.width.toFloat,
+      windowHeight - textureInfo.height.toFloat,
       textureInfo.width.toFloat,
       textureInfo.height.toFloat
     )
